# Make arkouda server configurable w.r.t. supported types