--- a/ServerModules.cfg
+++ b/ServerModules.cfg
@@ -1,0 +1,40 @@
+# Comment out modules using a `#` to exclude
+# a module from a build
+
+ArgSortMsg
+ArraySetopsMsg
+BroadcastMsg
+CastMsg
+ConcatenateMsg
+CSVMsg
+DataFrameIndexingMsg
+EfuncMsg
+EncodingMsg
+FlattenMsg
+HashMsg
+HDF5Msg
+HistogramMsg
+In1dMsg
+IndexingMsg
+JoinEqWithDTMsg
+KExtremeMsg
+LogMsg
+OperatorMsg
+ParquetMsg
+RandMsg
+ReductionMsg
+RegistrationMsg
+SegmentedMsg
+SequenceMsg
+SortMsg
+StatsMsg
+TimeClassMsg
+TransferMsg
+UniqueMsg
+
+# Add additional modules located outside
+# of the Arkouda src/ directory below.
+# For modules outside of the Arkouda src/
+# directory, make sure to add the complete
+# path to the module in the environment
+# variable ARKOUDA_SERVER_USER_MODULES.