Adds a new `serverConfig.json` file that contains options for specifying maximum array dimensionality (from https://github.com/Bears-R-Us/arkouda/pull/2829), as well as some new options to specify which scalar types should be supported by the server.

With the new options in place, Arkouda will be able to improve its [array_api](https://data-apis.org/array-api/latest/API_specification/data_types.html) compliance, where all of numpy's scalar data types are to be supported. This will also allow us to improve the "Arkouda as a numpy drop-in replacement" story when the server is compiled to support all the scalar types. When support for a given type is not desired, it will be ignored at compile time (using some `config param` values), leading to faster compilation and a smaller binary.

Other improvements:
* the `randIntMsg`, `castMsg`, `createMsg`, and `setMsg` procedures will now support all the numpy scalar types. This was implement as a proof-of-concept, with the intention of incrementally expanding type-configurability support in later PRs
* command dispatching now detects when an unrecognized command was not in the table because (1) the relevant command doesn't support ND arrays, or (2) the server was not compiled with a high enough maximum array dimensionality to support the command. In either of these cases, a more specific error message is returned to the client.